feat(library): add subtle bottom floating title pill to 1x1 folder cards (#1056)
